Move verilog_iofile class to a dedicated module file
	new file:   .gitignore
	modified:   verilog/__init__.py
	new file:   verilog/verilog_iofile.py
