+++ .gitignore
@@ -1,0 +1,5 @@
+*.swp
+*.swo
+*.pyc
+*~
+
